splitted navbar file into it own component files
