+++ material-ui-test/src/routes/home/components/navbar/navbar.styles.ts
@@ -1,0 +1,5 @@
+export const defaultNavbarStyles = {
+    bgcolor: "background.secondary",
+    position: "sticky",
+  };
+
